Modification entités, mise en place des commandes pour l'écriture

+++ Entities/Exercice.cs
@@ -9,7 +9,7 @@ namespace Taxes.Entities
     public class Exercice
     {
         [Key]
-        public int Anee_exercice { get; set; }
+        public short Anee_exercice { get; set; }
         public DateTime Date_echeance { get; set; }
         public DateTime Date_reglement_taxe { get; set; }
